smartpath-ai/backend/requirements.txt
# Core FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data processing and analysis
pandas==2.1.3
numpy==1.25.2

# AI and Machine Learning
google-generativeai==0.3.2
sentence-transformers==2.2.2
torch==2.1.1
transformers==4.35.2
scikit-learn==1.3.2

# PDF and document processing
PyMuPDF==1.23.8

# HTTP requests
requests==2.31.0

# Streamlit for frontend
streamlit==1.28.1

# Additional utilities
python-dotenv==1.0.0
pydantic==2.5.0
